+++ src/main/java/sg/supermarket_kata/MarketCatalogue.java
@@ -3,6 +3,7 @@
  */
 package sg.supermarket_kata;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -12,14 +13,14 @@ import java.util.Map;
  */
 public class MarketCatalogue {
 	private Map<String, Product> products = new HashMap<String, Product>();
-	private Map<String, Double> prices = new HashMap<String, Double>();
+	private Map<String, BigDecimal> prices = new HashMap<String, BigDecimal>();
 
-	public void addProduct(Product product, double price) {
+	public void addProduct(Product product, BigDecimal price) {
 		this.products.put(product.getName(), product);
 		this.prices.put(product.getName(), price);
 	}
 
-	public double getUnitPrice(Product p) {
+	public BigDecimal getUnitPrice(Product p) {
 		return this.prices.get(p.getName());
 	}
 }
